Use py2k selectors instead of comparisons
For some reason comparison selectors on packages result in faulty
re-renderings. Unclear as to why this happens at this point. However, we
can normally convert them into something equivalent as a workaround.
Hopefully we can address this so it isn't an issue in the future. In any
event, we use the effective workaround for now.